+++ controllers/add.js
@@ -1,0 +1,9 @@
+const { Contact } = require("../models/contactShema");
+
+const add = async (req, res) => {
+  const result = await Contact.create(req.body);
+
+  res.status(201).json({ status: "success", code: 201, data: result });
+};
+
+module.exports = add;
