+++ controllers/index.js
@@ -1,0 +1,17 @@
+const { controller } = require("../helpers");
+
+const getAll = require("./getAll");
+const getById = require("./getById");
+const add = require("./add");
+const deleteContact = require("./deleteContact");
+const update = require("./update");
+const updateFavorite = require("./updateFavorite");
+
+module.exports = {
+  getAll: controller(getAll),
+  getById: controller(getById),
+  add: controller(add),
+  deleteContact: controller(deleteContact),
+  update: controller(update),
+  updateFavorite: controller(updateFavorite),
+};
